+++ packages/cxalio-studio-tools/media_scout/appenv.py
@@ -8,7 +8,7 @@ class AppEnv(IAppEnvironment):
     def __init__(self):
         super().__init__()
         self.app_name = "MediaScout"
-        self.app_version = "0.8.0"
+        self.app_version = "0.8.6"
         self.output_console = Console()
         self.context: AppContext
 
@@ -25,7 +25,7 @@ class AppEnv(IAppEnvironment):
         return self.context.debug_mode
 
     def show_banner(self):
-        self.say(f"[blue]{self.app_name}[/] [yellow]v{self.app_version}[/]")
+        self.say(f"[cx.info]{self.app_name}[/] [cx.number]v{self.app_version}[/]")
 
 
 appenv = AppEnv()